Move relation into own file
